Get rid of ProcessorManagerProxy
