+++ sqs_workers/processor_mgr.py
@@ -11,56 +11,6 @@ from sqs_workers.utils import adv_bind_arguments
 logger = logging.getLogger(__name__)
 
 
-class ProcessorManagerProxy(object):
-
-    name = None  # type: str
-    env = None
-
-    def processor(
-        self,
-        job_name,
-        pass_context=False,
-        context_var=DEFAULT_CONTEXT_VAR,
-        backoff_policy=None,
-    ):
-        """
-        Decorator to assign processor to handle jobs with the name job_name
-        from the queue queue_name
-
-        Usage example:
-
-        queue = sqs.queue('q1')
-
-        @queue.processor('say_hello')
-        def say_hello(name):
-            print("Hello, " + name)
-
-        Then you can add messages to the queue by calling ".delay" attribute
-        of a newly created object:
-
-        >>> say_hello.delay(name='John')
-
-        Here, instead of calling function locally, it will be pushed to the
-        queue (essentially, mimics the non-blocking call of the function).
-
-        If you still want to call function locally, you can call
-
-        >>> say_hello(name='John')
-        """
-
-        def fn(processor):
-            return self.env.processors.connect(
-                self.name,
-                job_name,
-                processor,
-                pass_context,
-                context_var,
-                backoff_policy,
-            )
-
-        return fn
-
-
 class ProcessorManager(object):
     def __init__(
         self,
